Project 2 [CryptoInsight]/backend/requirements.txt
python-dotenv~=1.0.0
pymongo~=4.6.1
httpx~=0.25.2
colorama~=0.4.6
openai~=1.5.0
fastapi~=0.108.0
pydantic~=2.5.2
uvicorn~=0.25.0
passlib~=1.7.4
python-jose[cryptography]~=3.3.0
bcrypt~=4.1.1
python-multipart~=0.0.6
uuid6~=2023.5.2